Chamgement de types (incohérence) + mesure du temps pour l'algo Dijkstra

+++ LivrOnTime/src/main/java/util/tsp/TemplateTSP.java
@@ -13,7 +13,7 @@ public abstract class TemplateTSP implements TSP {
 		return tempsLimiteAtteint;
 	}
 	
-	public void chercheSolution(int tpsLimite, int nbSommets, long[][] cout, long[] duree,long[][] time){
+	public void chercheSolution(long tpsLimite, int nbSommets, long[][] cout, long[] duree,long[][] time){
 		
 		
 		tempsLimiteAtteint = false;
@@ -68,7 +68,7 @@ public abstract class TemplateTSP implements TSP {
 	 * @param tpsDebut : moment ou la resolution a commence
 	 * @param tpsLimite : limite de temps pour la resolution
 	 */	
-	 void branchAndBound(int sommetCrt, ArrayList<Integer> nonVus, ArrayList<Integer> vus, long coutVus, long[][] cout, long[] duree, long tpsDebut, int tpsLimite, long[][] time){
+	 void branchAndBound(int sommetCrt, ArrayList<Integer> nonVus, ArrayList<Integer> vus, long coutVus, long[][] cout, long[] duree, long tpsDebut, long tpsLimite, long[][] time){
 		 if (System.currentTimeMillis() - tpsDebut > tpsLimite){
 			 tempsLimiteAtteint = true;
 			 System.err.println("Temps limite atteint !");
